+++ ReleaseNotes.md
@@ -7,6 +7,8 @@ A. Start with the Quick Start guide : https://github.com/SteveGilham/altcover/wi
 # 3.5.5xx (Cerulean series release 8)
 * `altcover version` command line option to output the version of AltCover being used (also `Invoke-AltCover -Version`, MSBuild task `AltCover.GetVersion`, option `/p:AltCoverGetVersion`)
 * `--single` (`-Single`, `Single="bool"`, `/p:AltCoverSingle`) option to only sample the first visit to each point (per AppDomain), reducing intermediate file size and processing time.
+* `--linecover` (`-LineCover`, `LineCover="bool"`, `/p:AltCoverLineCover`) option to only only record line coverage in OpenCover format.
+* `--branchcover` (`-BranchCover`, `BranchCover="bool"`, `/p:AltCoverBranchCover`) option to only record branch coverage in OpenCover format.
 
 # 3.5.560 (Cerulean series release 7)
 * [BUGFIX] -- `dotnet test` integration : don't leave the imported FSharp.Core.dll in the binaries folder